feat: Add base BLoC event and state classes
This commit adds the base BLoC event and state classes to the project. These classes serve as the foundation for creating BLoC events and states in the application. They are abstract classes that provide a common structure and functionality for all BLoC events and states.

The following files were added or modified:
- lib/src/app/base/bloc/base_bloc_event.dart
- lib/src/app/base/bloc/base_bloc_state.dart
- lib/src/app/pages/register/bloc/register_event.dart
- lib/src/app/bloc/app_state.dart
- lib/src/app/pages/login/bloc/login_state.dart
- lib/src/app/pages/register/bloc/register_state.dart
- lib/src/app/pages/login/bloc/login_event.dart
- lib/src/data/interfaces/auth_service_interface.dart
- lib/src/app/bloc/app_event.dart
- lib/src/data/models/login_request.dart
- lib/src/app/pages/dashboard/dashboard_page.dart
- lib/src/domain/entities/user_entity.dart
- pubspec.yaml