feat: Add Result and ActiveFriendEntity classes
This commit adds the `Result` class and `ActiveFriendEntity` class to the project. The `Result` class is a generic class that represents the result of an operation, with a `success` variant and an `error` variant. The `ActiveFriendEntity` class represents an active friend, with properties such as `isActive`, `id`, `firstName`, `lastName`, and `email`.

The following files were added or modified:
- lib/src/shared/async_helper/result.dart
- lib/src/app/pages/dashboard/bloc/dashboard_state.dart
- lib/src/shared/async_helper/run_catching.dart
- lib/src/domain/entities/active_friend_entity.dart
- lib/src/domain/entities/friend_request_entity.dart
- lib/src/app/pages/dashboard/bloc/dashboard_event.dart
- lib/src/domain/entities/friend_request_entity.g.dart
- lib/src/app/navigation/middleware/auth_guard.dart
- lib/src/domain/entities/active_friend_entity.g.dart