feat: Update environment variables and add new URLs
This commit updates the environment variables in the `.env` file to include new URLs for the `PRESENCE_HUB` and `CONVERSATION_HUB`. It also adds the `PRESENCE_HUB_URL_KEY` and `CONVERSATION_HUB_URL_KEY` to the `.env` file. Additionally, the `BaseClient` class is updated to include a new method `currentUser()` for fetching the current user.

The following files were modified:
- .env
- devtools_options.yaml
- lib/main.dart
- lib/src/app/network/clients/base_client.dart
- lib/src/app/network/interceptors/error_interceptor.dart
- lib/src/domain/repositories/auth_service.dart
- lib/src/app/pages/chats/subviews/bloc/chat_state.dart
- lib/src/data/interfaces/auth_service_interface.dart
- lib/src/app/pages/chats/subviews/bloc/chat_event.dart
- lib/src/shared/extensions/datetime_extension.dart
- lib/src/domain/entities/user_entity.dart
- lib/src/domain/entities/active_friend_entity.dart
- lib/application_config.dart
- lib/src/shared/hubs/base_hub.dart