feat: Improve calculation of day difference in DateTime extension
Refactor the `getDay` method in the `TimeDifference` extension of the `DateTime` class to improve the calculation of the day difference. Instead of directly comparing the difference in days between the current date and the given date, the method now compares the difference between the current date and the given date at midnight. This ensures accurate calculation of the day difference, especially when dealing with time zones and daylight saving time changes.

Modified file:
- lib/src/shared/extensions/datetime_extension.dart